"Fix" reverse fitness sorting
Agents with the lowest fitness were being picked as survivors
rather than those with the highest fitness.
Strangely, fitness was still increasing.
This is because of the properties of the Iterated Prisoner's Dilemma.
However, fitness stayed low.
Fixing this bug somehow makes fitness fail to increase with evolution
for the IPD. Still trying to figure out why.